+++ lib/src/widgets/stories_list.dart
@@ -1,22 +1,38 @@
+import 'package:dashbook/src/widgets/icon.dart';
 import 'package:flutter/material.dart';
 import './side_bar_panel.dart';
 import './link.dart';
 import '../story.dart';
 
 typedef OnSelectChapter = Function(Chapter chapter);
+typedef OnBookmarkChapter = Function(String chapter);
 
 class StoriesList extends StatelessWidget {
   final List<Story> stories;
   final Chapter? selectedChapter;
   final OnSelectChapter onSelectChapter;
+  final String? currentBookmark;
+  final OnBookmarkChapter onBookmarkChapter;
+  final VoidCallback onClearBookmark;
   final VoidCallback onCancel;
 
   StoriesList({
     required this.stories,
+    required this.currentBookmark,
+    required this.onBookmarkChapter,
+    required this.onClearBookmark,
     required this.onSelectChapter,
     required this.onCancel,
     this.selectedChapter,
   });
+
+  void _pin(Chapter chapter) {
+    if (chapter.id == currentBookmark) {
+      onClearBookmark();
+    } else {
+      onBookmarkChapter(chapter.id);
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -44,28 +60,47 @@ class StoriesList extends StatelessWidget {
                 initiallyExpanded: true,
                 children: [
                   for (Chapter chapter in story.chapters)
-                    GestureDetector(
-                      behavior: HitTestBehavior.opaque,
-                      child: Align(
-                        alignment: Alignment.centerLeft,
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(horizontal: 16),
-                          child: Link(
-                            label: "  ${chapter.name}",
-                            textAlign: TextAlign.left,
-                            padding: EdgeInsets.zero,
-                            height: 20,
-                            textStyle: TextStyle(
-                              fontWeight: chapter.id == selectedChapter?.id
-                                  ? FontWeight.bold
-                                  : FontWeight.normal,
+                    Align(
+                      alignment: Alignment.centerLeft,
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 16),
+                        child: Row(
+                          children: [
+                            Expanded(
+                              child: GestureDetector(
+                                onTap: () {
+                                  onSelectChapter(chapter);
+                                },
+                                behavior: HitTestBehavior.opaque,
+                                child: Link(
+                                  label: "  ${chapter.name}",
+                                  textAlign: TextAlign.left,
+                                  padding: EdgeInsets.zero,
+                                  height: 20,
+                                  textStyle: TextStyle(
+                                    fontWeight:
+                                        chapter.id == selectedChapter?.id
+                                            ? FontWeight.bold
+                                            : FontWeight.normal,
+                                  ),
+                                ),
+                              ),
                             ),
-                          ),
+                            if (true)
+                              Opacity(
+                                opacity:
+                                    chapter.id == currentBookmark ? 1 : 0.05,
+                                child: DashbookIcon(
+                                  icon: Icons.bookmark,
+                                  onClick: () => _pin(chapter),
+                                  tooltip: chapter.id == currentBookmark
+                                      ? 'Remove this chapter'
+                                      : 'Bookmark this bookmark',
+                                ),
+                              ),
+                          ],
                         ),
                       ),
-                      onTap: () {
-                        onSelectChapter(chapter);
-                      },
                     ),
                   SizedBox(height: 10),
                 ],