+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - Adding control property to all properties
  - Large screens preview container share the space with the properties contaner for better visualization
  - Adding info property to chapters
+ - Adding bookmark handler for chapters
 
 ## [0.1.1]
  - Fixing null theme